Fix nvme-tcp module loading to be non-fatal
The nvme-tcp kernel module may not be available on all runner kernel
versions. Make modprobe failure non-fatal to allow tests to continue.

Error was: modprobe: FATAL: Module nvme-tcp not found in directory
/lib/modules/6.8.0-88-generic

This was accidentally introduced when simplifying the setup actions
for pristine VMs - we removed the error handling that made this
non-fatal.