fix: use correct TrueNAS API parameters for GetDatasetProperties
The pool.dataset.query 'extra' options were using 'properties: true' which
caused a TypeError in TrueNAS ('bool' object is not iterable) because
TrueNAS expects 'properties' to be a list of ZFS property names.

Changed to use the correct parameters:
- 'user_properties': true - to get user-defined ZFS properties
- 'flat': true - return flat list instead of tree
- 'retrieve_children': false - don't retrieve child datasets

This fixes deleteStrategy=retain not working because GetDatasetProperties
was failing and the code was falling back to the default 'delete' strategy.